[Add] - Added regist course function
