+++ StudentManagement/main.c
@@ -6,9 +6,10 @@ int main()
 {
 	int studentNumber = 0;
 	int subjectNumber = 0;
-    STUDENT student[100];
+
+	COURSE *course;
     
 	SystemSet();
-	CourseMenu();
+	CourseMenu(&course, &subjectNumber);
     return 0;
 }
